+++ src/lib/Components/Consignments/Screens/SelectFromPhotoLibrary.tsx
@@ -8,17 +8,8 @@ import { BodyText as P } from "../Typography"
 
 import { triggerCamera } from "lib/NativeModules/triggerCamera"
 
-import {
-  Alert,
-  CameraRoll,
-  Dimensions,
-  Linking,
-  NativeModules,
-  Route,
-  ScrollView,
-  View,
-  ViewProperties,
-} from "react-native"
+import CameraRoll from "@react-native-community/cameraroll"
+import { Alert, Dimensions, Linking, NativeModules, Route, ScrollView, View, ViewProperties } from "react-native"
 import NavigatorIOS from "react-native-navigator-ios"
 
 import { ConsignmentSetup } from "../index"
@@ -78,6 +69,7 @@ export default class SelectFromPhotoLibrary extends React.Component<Props, State
     const fetchParams: any = {
       first: 20,
       assetType: "Photos",
+      groupTypes: "All",
     }
 
     if (this.state.lastCursor) {
@@ -146,6 +138,7 @@ export default class SelectFromPhotoLibrary extends React.Component<Props, State
           const fetchParams: any = {
             first: 1,
             assetType: "Photos",
+            groupTypes: "All",
           }
 
           this.getCameraRollPhotos(fetchParams).then(photos => {