+++ src/lib/Components/Consignments/Screens/__tests__/Overview-uploading-tests.tsx
@@ -5,6 +5,8 @@ AsyncStorage.removeItem = jest.fn()
 
 import Overview from "../Overview"
 
+jest.mock("@react-native-community/cameraroll", () => jest.fn())
+
 jest.mock("../../Submission/update", () => jest.fn())
 import updateSubmission from "../../Submission/update"