+++ src/lib/Scenes/Artwork/Artwork.tsx
@@ -1,10 +1,14 @@
+import { Flex } from "@artsy/palette"
 import { Artwork_artwork } from "__generated__/Artwork_artwork.graphql"
 import { ArtworkQuery } from "__generated__/ArtworkQuery.graphql"
+import Separator from "lib/Components/Separator"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
 import React from "react"
-import { Text, View } from "react-native"
+import { View } from "react-native"
 import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
+import { ArtworkActionsFragmentContainer as ArtworkActions } from "./Components/ArtworkActions"
+import { ArtworkAvailabilityFragmentContainer as ArtworkAvailability } from "./Components/ArtworkAvailability"
 import { ArtworkTombstoneFragmentContainer as ArtworkTombstone } from "./Components/ArtworkTombstone"
 
 interface Props {
@@ -15,9 +19,14 @@ export class Artwork extends React.Component<Props> {
   render() {
     const { artwork } = this.props
     return (
-      <View style={{ backgroundColor: "white" }}>
-        <Text>{this.props.artwork.title}</Text>
-        <ArtworkTombstone artwork={artwork} />
+      <View>
+        <Flex width="100%" style={{ backgroundColor: "gray" }} height={340} />
+        <Flex alignItems="center" mt={2}>
+          <ArtworkTombstone artwork={artwork} />
+          <ArtworkActions artwork={artwork} />
+          <Separator />
+          <ArtworkAvailability artwork={artwork} />
+        </Flex>
       </View>
     )
   }
@@ -26,8 +35,9 @@ export class Artwork extends React.Component<Props> {
 export const ArtworkContainer = createFragmentContainer(Artwork, {
   artwork: graphql`
     fragment Artwork_artwork on Artwork {
-      title
       ...ArtworkTombstone_artwork
+      ...ArtworkActions_artwork
+      ...ArtworkAvailability_artwork
     }
   `,
 })