+++ src/lib/Scenes/Artwork/__stories__/Artwork.story.tsx
@@ -4,7 +4,7 @@ import React from "react"
 import { ArtworkRenderer } from "../Artwork"
 import { ArtworkTombstone } from "../Components/ArtworkTombstone"
 
-storiesOf("Artwork")
+storiesOf("Artwork/Screens")
   .add("Institution", () => {
     return <ArtworkRenderer artworkID="pablo-picasso-le-reve-the-dream" />
   })